[DRAFT] unison stereo spread
only works for basic waves (and wavetables) yet, not samples nor FM/RM.
I'm gonna try refactor the mono vs stereo rendering logic, to
not duplicate the same panning logic for samples, FM and ringmod.

For now the shortcut is to press SELECT inside the unison number menu,
but we could also take the free spot above it (right of decimation).