feature(sysex): better screen updates with less CPU
Instead of having the client poll the screen with regular
intervals, this instead has the deluge push out updates
whenever the screen changes. This should dramatically reduce
the CPU overhead of this feature when there is no changes to the
OLED display.

Implement throttling by checking the fullness of the sysex send
buffer. This naturally reduces the refresh rate as needed when
using serial MIDI, or some USB device which reading too slow

he client still has to continuously ask the deluge for updates,
to stop sending after the client disconnects. This can be done
with a much lower rate, one short message per second. The client
can also force a full update so a dropped delta message does
not cause the screen to be corrupt forever.